requirements.txt
twscrape
pandas
requests
fredapi
python-dotenv
tradingeconomics
yfinance
matplotlib
statsmodels
dbnomics